requirements.txt
# EdgeCAAI-Net dependencies
torch>=2.1.0
torchaudio>=2.1.0
librosa>=0.10.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
ptflops>=0.7.0
soundfile>=0.12.0
